Correction ejercicio005
